f64 -> f32
Turns out f64 is *really slow* on the wasm32 target. We don't actually need the extra precision for anything, so kill it with fire.

+++ src/apu/audio_channel.rs
@@ -6,9 +6,9 @@ use super::RingBuffer;
 
 #[derive(Clone)]
 pub enum PlaybackRate {
-    FundamentalFrequency { frequency: f64 },
+    FundamentalFrequency { frequency: f32 },
     LfsrRate { index: usize, max: usize },
-    SampleRate { frequency: f64 },
+    SampleRate { frequency: f32 },
 }
 
 #[derive(Clone)]
@@ -41,12 +41,12 @@ pub trait AudioChannelState {
     fn rate(&self) -> PlaybackRate { return PlaybackRate::SampleRate{frequency: 0.0}; }
     fn volume(&self) -> Option<Volume> {return None}
     fn timbre(&self) -> Option<Timbre> {return None}
-    fn amplitude(&self) -> f64 {
+    fn amplitude(&self) -> f32 {
         /* pre-mixed volume, allows chips using non-linear mixing to tailor this value.
            results should be based on 2A03 pulse, where 1.0 corresponds to 0xF */
         if !self.playing() {return 0.0}
         match self.volume() {
-            Some(Volume::VolumeIndex{index, max}) => {return index as f64 / (max + 1) as f64},
+            Some(Volume::VolumeIndex{index, max}) => {return index as f32 / (max + 1) as f32},
             None => {return 1.0}
         }
     }
