+++ src/apu/pulse.rs
@@ -156,7 +156,7 @@ impl AudioChannelState for PulseChannelState {
     }
 
     fn record_current_output(&mut self) {
-        self.debug_filter.consume(self.output() as f64);
+        self.debug_filter.consume(self.output() as f32);
         self.output_buffer.push((self.debug_filter.output() * -4.0) as i16);
         self.edge_buffer.push(self.last_edge as i16);
         self.last_edge = false;
@@ -191,7 +191,7 @@ impl AudioChannelState for PulseChannelState {
     }
 
     fn rate(&self) -> PlaybackRate {
-        let frequency = self.cpu_clock_rate as f64 / (16.0 * (self.period_initial as f64 + 1.0));
+        let frequency = self.cpu_clock_rate as f32 / (16.0 * (self.period_initial as f32 + 1.0));
         return PlaybackRate::FundamentalFrequency {frequency: frequency};
     }
